pyproject.toml
[project]
name = "gpt-models"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "satoooh", email = "bighope.lumiere0916@gmail.com" }
]
dependencies = [
    "python-dotenv>=1.0.1",
    "openai>=1.30.1",
    "matplotlib>=3.9.0",
    "pandas>=2.2.2",
    "statsmodels>=0.14.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/gpt_models"]
